Some refactor with getting addresses list (#35)
