+++ src/api/banned_addresses_router.py
@@ -1,8 +1,5 @@
-import datetime
-import logging
 from dataclasses import asdict
 from typing import Annotated
-from typing import Any
 from typing import Optional
 
 from fastapi import APIRouter
@@ -10,13 +7,13 @@ from fastapi import BackgroundTasks
 from fastapi import Depends
 from fastapi.security import HTTPAuthorizationCredentials
 
+from src.api.di.banned_addresses_routines import get_banned_addresses
 from src.core.settings import ACTIVE_USAGE_INFO
 from src.core.settings import BACKGROUND_ADD_RECORDS
 from src.core.settings import BACKGROUND_DELETE_RECORDS
 from src.core.settings import HISTORY_USAGE_INFO
 from src.db.adapters.redis_set_db_entity_adapter import RedisSetDbEntityAdapter
 from src.db.adapters.set_db_entity_str_adapter import SetDbEntityStrAdapterIpAddress
-from src.db.adapters.set_db_entity_str_adapter import SetDbEntityStrAdapterIpNetwork
 from src.db.storages.redis_db import RedisAsyncio
 from src.db.storages.redis_db import redis_client
 from src.models.query_params_models import CommonQueryParams
@@ -26,10 +23,7 @@ from src.schemas.common_response_schemas import AddResponseSchema
 from src.schemas.common_response_schemas import CountResponseSchema
 from src.schemas.common_response_schemas import DeleteResponseSchema
 from src.schemas.usage_schemas import ActionType
-from src.service.addresses_db_service import AllowedAddressesSetDBService
 from src.service.addresses_db_service import BlackListAddressesSetDBService
-from src.service.networks_db_service import AllowedNetworksSetDBService
-from src.service.process_banned_ips import without_allowed_ips
 from src.tasks.celery_tasks import celery_update_history_task
 from src.tasks.celery_tasks import celery_update_usage_info_task
 from src.tasks.history_update_bg_task import update_history_bg_task_ns
@@ -39,39 +33,6 @@ from .http_auth_wrapper import get_proc_auth_checker
 
 api_router = APIRouter()
 banned_addresses_auth_check = get_proc_auth_checker(need_admin_permission=False)
-
-
-async def get_banned_addresses(redis_client_obj: RedisAsyncio, query_params: dict[str, Any]) -> IpV4AddressList:
-    # TODO transform fetching logic to generator
-    start_moment = datetime.datetime.now()
-    try:
-        logging.debug('Starting blacklist query execution')
-        service_obj = BlackListAddressesSetDBService(
-            SetDbEntityStrAdapterIpAddress(RedisSetDbEntityAdapter(redis_client_obj))
-        )
-        filter_records = query_params.pop('filter_records')
-        # TODO There's no need to get all records. We can fetch them for further filtering and return them without
-        # memory utilization and get the right records count
-        # At the moment some records might be truncated with following filtering
-        banned_records = await service_obj.get_records(**query_params)
-        if filter_records:
-            # filter by allowed IPs
-            allowed_service_obj = AllowedAddressesSetDBService(
-                SetDbEntityStrAdapterIpAddress(RedisSetDbEntityAdapter(redis_client_obj))
-            )
-            allowed_records = await allowed_service_obj.get_records(all_records=True)
-            allowed_net_service_obj = AllowedNetworksSetDBService(
-                SetDbEntityStrAdapterIpNetwork(RedisSetDbEntityAdapter(redis_client_obj))
-            )
-            allowed_networks = await allowed_net_service_obj.get_records(all_records=True)
-            return [x async for x in without_allowed_ips(banned_records, allowed_records, allowed_networks)]
-        return banned_records
-    finally:
-        duration = datetime.datetime.now() - start_moment
-        logging.debug(
-            'Finished blacklist query execution, elapsed %s milliseconds',
-            duration.seconds + duration.microseconds / 1000,
-        )
 
 
 @api_router.get(
